requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
django-pytest==0.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
